добавил новый эндпоинт для получения последних новостей. Создал обертку для пагинации
