+++ src/components/LatestNews/index.jsx
@@ -1,10 +1,13 @@
+import { getLatestNews } from "../../api/apiNews";
 import BannersList from "../BannersList";
 import styles from "./styles.module.css";
+import { useFetch } from "../../helpers/hooks/useFetch";
 
-const LatestNews = ({ banners, isLoading }) => {
+const LatestNews = () => {
+    const { data, isLoading } = useFetch(getLatestNews);
     return (
         <section className={styles.section}>
-            <BannersList banners={banners} isLoading={isLoading} />
+            <BannersList banners={data?.news} isLoading={isLoading} />
         </section>
     );
 };
