+++ src/components/NewsByFilters/index.jsx
@@ -1,10 +1,28 @@
-import { TOTAL_PAGES } from "../../constants";
-import Pagination from "../Pagination";
+import { PAGE_SIZE, TOTAL_PAGES } from "../../constants";
 import styles from "./styles.module.css";
 import NewsList from "../NewsList";
 import NewsFilters from "../NewsFilters";
+import { getNews } from "../../api/apiNews";
+import { useDebounce } from "../../helpers/hooks/useDebounce";
+import { useFetch } from "../../helpers/hooks/useFetch";
+import { useFilters } from "../../helpers/hooks/useFilters";
+import PaginationWrapper from "../PaginationWrapper";
 
-const NewsByFilters = ({ filters, changeFilter, isLoading, news }) => {
+const NewsByFilters = () => {
+    const { filters, changeFilter } = useFilters({
+        page_number: 1,
+        page_size: PAGE_SIZE,
+        category: null,
+        keywords: "",
+    });
+
+    const debouncedKeywords = useDebounce(filters.keywords, 1500);
+
+    const { data, isLoading } = useFetch(getNews, {
+        ...filters,
+        keywords: debouncedKeywords,
+    });
+
     const handleNextPage = () => {
         if (filters.page_number < TOTAL_PAGES) {
             changeFilter("page_number", filters.page_number + 1);
@@ -20,26 +38,21 @@ const NewsByFilters = ({ filters, changeFilter, isLoading, news }) => {
     const handlePageClick = (pageNumber) => {
         changeFilter("page_number", pageNumber);
     };
+
     return (
         <section className={styles.section}>
             <NewsFilters filters={filters} changeFilter={changeFilter} />
-            <Pagination
+            <PaginationWrapper
+                top
+                bottom
                 handlePreviousPage={handlePreviousPage}
                 handleNextPage={handleNextPage}
                 handlePageClick={handlePageClick}
                 totalPages={TOTAL_PAGES}
                 currentPage={filters.page_number}
-            />
-
-            <NewsList isLoading={isLoading} news={news} />
-
-            <Pagination
-                handlePreviousPage={handlePreviousPage}
-                handleNextPage={handleNextPage}
-                handlePageClick={handlePageClick}
-                totalPages={TOTAL_PAGES}
-                currentPage={filters.page_number}
-            />
+            >
+                <NewsList isLoading={isLoading} news={data?.news} />
+            </PaginationWrapper>
         </section>
     );
 };
